Add EmailAuthorization_AuthorizedGroups
This adds the capability of adding one or more groups that are
authorized regardless of email address. Among other things, this
helps to bootstrap the extension when it is initially installed:
'sysop' can be set to be an authorized group so an initial sysop
user can gain access to the wiki.

Change-Id: I9278839cb5ba68ef31e8cd222e6fd1bfbbae7564